refactor(host): repoint doc routes + homepages to _chrome adapter (#2427)
The 4 doc-route entrypoints now import renderDocPage from ../lib/_chrome
(was _doc-page-renderer). The 2 homepages import HeaderWithDefaults /
FooterWithDefaults / HeadWithDefaults / composeMetaTitle from ./lib/_chrome
(was the per-component _*-with-defaults shims); their BodyEndIslands import
stays at the kept _body-end-islands seam. Import-line rewiring only — these
are the same factories called with the same chromeCtx, so rendered output is
byte-identical (verified by parity --verify).
